* Make gen_leader handle the case where only one Candidate is supplied * Make the Rakefile handle modules=mod1,mod2 to determine which modules to run tests for (so you can test one or 2 modules by themselves)
git-svn-id: http://svn.dev/cpx/cpxerl@2528 31352e9c-24da-da11-96a7-0090fb0792c6

+++ Rakefile
@@ -52,21 +52,17 @@ end
 # this almost works, doesn't handle module dependancies though :(
 rule ".txt" => ["%{coverage,debug_ebin}X.beam", 'debug_ebin/test_coverage.beam'] do |t|
 	mod = File.basename(t.source, '.beam')
+	if ENV['modules'] and not ENV['modules'].split(',').include? mod
+		puts "skipping tests for #{mod}"
+		next
+	end
+
 	test_output = `erl -pa debug_ebin -sname testpx -s test_coverage start #{mod} -run init stop`
 	if /(All \d+ tests successful|There were no tests to run|This module does not provide a test\(\) function)/ =~ test_output
 		puts "  #{mod.ljust(@maxwidth - 1)} : #{$1}"
 	else
 		puts test_output.split("\n")[1..-1].map{|x| x.include?('1>') ? x.gsub(/\([a-zA-Z0-9\-@]+\)1>/, '') : x}.join("\n")
 		File.delete(t.to_s)
-	#else # /=ERROR REPORT===/ =~ test_output
-		#puts test_output
-		#File.delete(t.to_s)
-	#elsif /=INFO REPORT===/ =~ test_output
-		#puts test_output
-		#File.delete(t.to_s)
-	#else
-		#test_output[/1>\s*(.*)\n/]
-		#puts "  #{mod.ljust(@maxwidth - 1)} : #{$1}"
 	end
 end
